# docs/api-swagger.yaml
openapi: 3.0.3
info:
  title: OPDA Dashboard API
  version: 0.1.0
  description: >
    OpenAPI stubs for OPDA Operations Dashboard (read-first).
    All endpoints require Bearer JWT unless stated otherwise.
servers:
  - url: http://localhost:8088
    description: Local (via NGINX)
...
security:
  - BearerAuth: []
paths:
...
  /api/incidents:
    get:
      tags: [incidents]
      summary: List incidents
      parameters:
        - $ref: "#/components/parameters/StatusParam"
        - $ref: "#/components/parameters/SeverityParam"
        - $ref: "#/components/parameters/AppCodeParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: Incident list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/incidents/{incident_id}:
    get:
      tags: [incidents]
      summary: Get incident detail
      parameters:
        - $ref: "#/components/parameters/IncidentIdPath"
      responses:
        "200":
          description: Incident detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized (missing/invalid token)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: Not authenticated
    Forbidden:
      description: Forbidden (insufficient permission)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: Not Found
  parameters:
    LimitParam:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    OffsetParam:
      name: offset
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    StatusParam:
      name: status
      in: query
      required: false
      schema: { type: string, example: OPEN }
    SeverityParam:
      name: severity
      in: query
      required: false
      schema: { type: string, example: P1 }
    AppCodeParam:
      name: app_code
      in: query
      required: false
      schema: { type: string, example: RTGS }
    IncidentIdPath:
      name: incident_id
      in: path
      required: true
      schema: { type: string, example: INC-20260106-0001 }
...
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
      required: [detail]

    # ---------- Summary ----------
...
    IncidentListResponse:
      type: object
      required: [items, limit, offset, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/IncidentItem"
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer, minimum: 0 }
    IncidentItem:
      type: object
      required: [id, severity, status, title, app_code, opened_at, sla_state]
      properties:
        id: { type: string, example: INC-20260106-0001 }
        severity: { type: string, example: P1 }
        status: { type: string, example: OPEN }
        title: { type: string }
        app_code: { type: string, example: RTGS }
        owner_team: { type: string, example: OpDA }
        opened_at:
          type: string
          format: date-time
        sla_due_at:
          type: string
          format: date-time
          nullable: true
        sla_state:
          type: string
          enum: [ON_TRACK, AT_RISK, BREACHED]
    IncidentDetail:
      allOf:
        - $ref: "#/components/schemas/IncidentItem"
        - type: object
          properties:
            description: { type: string }
            linked_workorders:
              type: array
              items: { type: string }
            events:
              type: array
              items:
                $ref: "#/components/schemas/IncidentEvent"
    IncidentEvent:
      type: object
      required: [at, type, by]
      properties:
        at:
          type: string
          format: date-time
        type:
          type: string
          example: ACK
        by:
          type: string
          example: coco

    # ---------- Workorders ----------
...